refact: Migrate `panel.user` to TypeScript
